VERSION 0.5.0 - alpha users must rebuild and restart!
Whew. This is a big one. More of a marathon than a sprint.

First, four big things:

1) This version breaks backward compatibility with all prior versions.
It's in alpha, I can do that.

2) The port has changed from 8993 to 9993 to mark this change. Probably
not necessary but why not? Also 8993/UDP turned out to be used by some
enterprise LDAP thingy, which doesn't matter much either but again why
not?

3) This version, unlike previous versions, does NOT auto-join the Earth
network. Soon there will be more than one net, and not everyone is going
to want to get dumped on a flat global LAN right out of the box. To
join Earth use the command line interface:

sudo zerotier-cli join bc8f9a8ee3000001

4) Finally, you will get a different IP on Earth. The whole cryptosystem
has changed and we're not going to bother with continuity issues in
alpha testing.

So what's changed? See the blog:

http://blog.zerotier.com/post/62991430345/alpha-zerotier-one-network-is-down-briefly

The net should be up shortly after this commit. If there are any issues,
0.5.0 will be rapidly followed by 0.5.1. :-) Otherwise the next sprint
will be finishing up support for private networks. Then it's off to the
races with BETA, then Windows. (Decided to move into beta before Windows
in all likelihood.)