VERSION 0.6.0 BETA: please upgrade!
Version 0.6.0 marks the transition of ZeroTier One from ALPHA to BETA.
Major updates to the web site and binary packages for MacOS and Linux
are coming soon, followed by Windows soon thereafter.

This version contains a number of changes including:

 * Speed improvements to encryption
 * A new much-improved identity algorithm, which unfortunately requires an
   identity regeneration. This should happen automatically, and should be
   the last time for a good long while assuming there's nothing wrong with
   what's here.
 * Cleaned up the Network::Config mess in the code, factored out Config
   into its own NetworkConfig class.
 * Lots of work to support private networks, which are still in testing.
   Concurrent with the web site update will be another minor release to
   include any fixes there.
 * Some changes to the protocol for better future-proofing.
 * Netconf support for ARP caching parameters configurable on per-network
   basis.

You must update to stay connected to the network; this version will not
talk to 0.5.0. After this, I'm going to be much more reluctant to make
incompatible changes.