Support DragonFly BSD
DragonFly BSD is a fork of FreeBSD 4.8 in 2003. They share many
similarities, so this patch makes ZeroTierOne build and work in the
similar way as on FreeBSD, except for the following two differences:

* DragonFly BSD doesn't have getifmaddrs(3).
* DragonFly BSD's tap(4) interface doesn't support to set MTU.

I have tested that ZeroTierOne builds and works fine on the latest
DragonFly BSD (master version as of 2019-09-17):

DragonFly v5.7.0.531.gb7d3e1-DEVELOPMENT #0: Tue Sep 17 16:13:34 UTC 2019

Meanwhile, define ZT_DEBUG_TRACE in make-bsd.mk when build the debug
version.