Windows interface and route metrics
5 is the lowest metrics an interface will get based
on Windows Automatic Metrics as far as I know.
https://docs.microsoft.com/en-us/troubleshoot/windows-server/networking/automatic-metric-for-ipv4-routes

If we set our metric to 3, there's a chance Windows will
use it for broadcast traffic, and things like game discovery
will work. If the user doesn't want that, they can leave the
zerotier network or manually set the metric on their
physical interface to < 3.

For routes with "via"'s we put the metric high,
in case it overlaps with a physical route to
the same subnet.

These could be plumbed through and set via the controller or
local.conf someday, istead of being hardcoded magic numbers.