Fix for missing dependencies in Docker image
When building my own Docker image of ZeroTier with the latest version 1.8.6, I noticed it failed due to missing dependencies related to SSL that the Dockerfile.release did not take into account. In my proposed solution to fix this error, I've decided to make use of the APT package manager instead of the DPKG tool to install the DEB package, thus resolving the dependencies automatically.